third_party/WebKit/LayoutTests/media/encrypted-media/encrypted-media-v2-events-expected.txt

EVENT(message)
This is a testharness.js-based test.
PASS Verify v2 events. 
Harness: the test ran to completion.
